Categories tabs

+++ app/src/main/java/dev/voleum/speedruncom/model/Times.kt
@@ -1,10 +1,10 @@
 package dev.voleum.speedruncom.model
 
 data class Times(val primary: String,
-                 val primary_t: Int,
+                 val primary_t: Double,
                  val realtime: String,
-                 val realtime_t: Int,
+                 val realtime_t: Double,
                  val realtime_noloads: String,
-                 val realtime_noloads_t: Int,
+                 val realtime_noloads_t: Double,
                  val ingame: String,
-                 val ingame_t: Int)
+                 val ingame_t: Double)
